Menu de configuracion de sitio con muchas funciones
- Se puede subir y  cambiar la imagen de fondo
- Se puede cambiar el color de tema
- Las dos anteriores config se guardan por navegador

- Se puede borrar la base de datos con un boton
- Se puede rellenar la base de datos con un menu

+++ mangaback/backendAPIs/mainAPI/dbConnector.py
@@ -7,7 +7,7 @@ import requests
 import math
 
 # Local storage folder
-folder = 'mangaDB' + os.sep
+folder = 'mangaDB'
 
 # Create connection
 # SQLITE only allows a single thread (since it's only one file), so we inform it that we want to use it as an import
@@ -60,14 +60,15 @@ def deleteDatabase():
     recreateDatabase()
     # Clear storage folder
     for filename in os.listdir(folder):
-        file_path = os.path.join(folder, filename)
-        try:
-            if os.path.isfile(file_path) or os.path.islink(file_path):
-                os.unlink(file_path)
-            elif os.path.isdir(file_path):
-                shutil.rmtree(file_path)
-        except Exception as e:
-            print('Failed to delete %s. Reason: %s' % (file_path, e))
+        if(filename != '.gitignore'):
+            file_path = os.path.join(folder, filename)
+            try:
+                if os.path.isfile(file_path) or os.path.islink(file_path):
+                    os.unlink(file_path)
+                elif os.path.isdir(file_path):
+                    shutil.rmtree(file_path)
+            except Exception as e:
+                print('Failed to delete %s. Reason: %s' % (file_path, e))
 
 def formatAsManga(data):
     """
@@ -114,7 +115,7 @@ def convertImagesToLocal(manga):
     res.raise_for_status()
     imageBytes = res.content
     unique_filename = str(uuid.uuid4()) + '.jpg'
-    with open(folder + unique_filename, 'wb') as f:
+    with open(folder + os.sep + unique_filename, 'wb') as f:
         f.write(imageBytes)
     manga['img']= '/mangaDB/' + unique_filename
     # Characters images
@@ -125,7 +126,7 @@ def convertImagesToLocal(manga):
         res.raise_for_status()
         imageBytes = res.content
         unique_filename = uuid.uuid4().hex + '.jpg'
-        with open(folder + unique_filename, 'wb') as f:
+        with open(folder + os.sep + unique_filename, 'wb') as f:
             f.write(imageBytes)
         character['image'] = '/mangaDB/' + unique_filename
         newcharacters.append(character)
@@ -286,6 +287,12 @@ def getLocalDBMeta():
     dbSize = convert_size(dbSize)
     totalMangas = cursor.execute("SELECT COUNT() FROM manga").fetchone()
     return {'filesSize':filesSize,'dbSize':dbSize,'totalMangas':totalMangas}
+
+def uploadFile(file):
+    ext = '.'+file.filename.rsplit('.', 1)[1].lower()
+    unique_filename = uuid.uuid4().hex + ext
+    file.save(folder + os.sep + unique_filename)
+    return '/'+folder+'/'+unique_filename
 
 if __name__ == '__main__':
     mangas = getMangas()
